vim: cmp - use UltiSnips and some other sources
I didn't manage to get the mappings for luasnip to work as I wanted yet.
The last issue in date is that if the completion menu is open I can't
jump to the next placeholder, even without selecting anything (but maybe
it was before I removed select = true ?).

I don't want to waste more time on that right now, since I already have
a lot of snippets that I don't want to gave up until I rewrote them in
Lua, I still need UltiSnips. So I will use it in place of luasnip for
now.