vim: use Space as Leader 🥴
I thought about it for a long time and never did it.
Mainly because I'm so used to the `,` that I don't want to have to
forget and learn new habits.
But also because of some insert mappings that would be problematic with
<Space>.

I finally decided to keep the `,` for the insert mappings and start my
journey of learning for other mappings 🤣