vim: ultisnips for snippet engine & luasnip for lsp
I can't manage to make UltiSnips work for LSP snippets.
It expands properly, I can jump but as soon as I start typing it stops
working.
I wonder if it's not ultisnips which receive some kind of event and
discard the snippet as being done.

So I kept luasnip which works great for LSP snippets with cmp and I kept
UltiSnips to not have to rewrite all the snippets that I use right now.
Especially since I didn't manage to create snippet like I want with
luasnip.