vim: use UltiSnips with my fork
I have a temporary fix for handling floating windows.
It happens that it was the last thing preventing me to keep using
UltiSnips like I'm use to.
When a floating window was open the snippet was cleared, I prevented it
from happening in insert mode to solve it for now.