nvim: improvement of the new LSP config
It's a messy commit becaused I tried a bunch of things at once and it
was a bit hard to split the diffs.
It fixes things like using LSP snippets with a fork of Ultisnips.
Update some PHP snippets
Use my fork of argwrap until my PR is merged.
etc.