nvim(folds): try to fix folds interaction with lspsaga
Diagnostics are not readable when folds are enabled.
So I tried to disable them by default because I don't have those kinds
of diagnostics in PHP.
And I updated the PHP setup to enable them at the correct level from the
start.

Will see if it's enought or not for my current use cases, the issue is
that fold seems to be a window configuration not a buffer one. So it's
not possible to decide based on the filetype or something like that.
Might require to do it in lspsaga directly when we open/close a float
but then if the float is part of the main window it might disable folds
for the main window as well ?