nvim(lsp): clean up configuration
Between `lspconfig` to help configuring the servers, `mason` to install the
servers, `neodev` to override `lspconfig` for `lua_lsp`...
It was becoming quiet a mess so I tried to split everything in a
dedicated folder so that it's a bit easier to navigate all this
configuration and making sure in which order what is configured.

That's why I don't rely on the ordering of plugins provided by packer
but instead I manage in which order I load the configuration.