nvim(fugitive): put it back in place until more neogit improvements
I kind of like how Neogit is designed, it's feels "nice".
But it's ugly, not convenient, missing features (mainly about moving
quickly in the status buffer), etc.
So I'll keep it and try to get use to it while working on some
improvements but I keep fugitive close by to not be blocked if I find a
new limitation to Neogit