nvim(scrollbar): remove it, nice attempt but more annoying than useful
I liked the idea to have an overview of things going wrong in the
buffer.
But it's not stable enough yet and not motivated enough to try to help.
It's not updated when not currently in the buffer, seems to even update
only on write.
You can't disable it for some buffers but not others.

So overall I'll wait a bit and check later if there was any improvement,
as long as I have the information in my status bar (diagnostics numbers,
failing tests) I'm good without it.