nvim(noice): try it out
There is some nice little things, kind of dressing but more "general".
I'm not convinced about `notify` though, the windows are always in the
way.
Having all result of commands in there is going to be pretty weird too,
hence the mapping to show them in a popup with `<S-Enter>`