nvim(lsp): replace gq mapping by lsp formatting
Couldn't make it conditional to the capabilities because intelephense is
registering the capabilities after the "init" response.
During the init, it will say that it doesn't handle it (which is what we
have access to when attaching a buffer to a client).
And only later it will dynamically register. So, until I find a way to
be notified on dynamic registration I enable it by default because I
never use this mapping anyway.