chore: a bunch of things because of bad organization
I did a bunch of change while not being willing to stop what I was doing
the time to comment, explain, commit
So I don't remember why I have most of them. Since they are here for a
while now, I keep them anyway and will try to be more diligent in the
future.