sxhkd(screenshot): use flameshot to take screenshot
I had issues with the `scrot` + `xclip` setup.
Which I also encountered when trying `maim` instead of `scrot`.
Probably came from `xlcip`, when taking big screenshots applications
like slack were going crazy when I tried to paste.
They were unresponsive until I copy something else, at which point the
new "copy thing" was pasted in slack.

At the end, flameshot still seems pretty lightweight and provide the
possibility to annotate screenshots on the flight. It can be really
useful so I kind of like it.

Since there was a bit of complexity to manage to configure all the use
cases I wanted, like selecting a monitor because flameshot and xrandr
don't use the same indexing, I write a little local script.

I took the opportunity to play with claude.ai and see what it could do.
This is why the script might be a bit "much" :)