git(hooks): copy CLAUDE.local.md when creating a new worktree
Since CLAUDE.local.md is not part of the git repository, when creating a
new worktree I loose it.
This hook allows me to copy it.
It's not a great idea but I wanted to play with hooks.
It has a lot of flaws, for instance the files are not kept aligned. So
you can have things added to it in one worktree but not accessible in
another.
It would probably be best to work with all Claude Code related files in
the parent directory.