nvim(lazy): moved to folder structure
I was thinking about it for a while but was "concerned" by the overhead
of having to read from the disk.
I saw that it's now the recommended way to do it with lazy anyway.
Plus, moving as much plugins as possible to be lazy loaded already
greatly improved the startup time.