nvim(comments): go back to tpope plugin instead of native NeoVim
I got too used to it, it's annoying to have a new behavior now.
When I comment multiple paragraphs, nvim will also comment the empty
lines. That's annoying because I often comment a bunch of stuff as one
and then uncomment one at a time with `gcuip` or alike.

There was a few other things, I think, since the main reason I switched
was to limits the dependencies on vim plugins and not because of
additional features.
I'll wait and see if the native comment get's more configurable before
switching back.