nvim(php): fix override of commentstring
I miss used the ftplugins, they are not made to override what the system
one is doing.
For this I MUST use after/ftplugin.
Which is why my commentstring override was not taken into consideration.

I also removed the one for gitcommit because it was defining what is
already the default behavior.