nvim(treesitter): fix loading config
Apparently the way to load the config changes, we do `require('nvim-treesitter.configs').setup(opts)`
I'm not sure lazy is gonna have a specific heuristic for that so I
overrode the `config` in order to tell it how to load the `opts`.
Seems to work ;)

I also tried to not install every parser anymore.
I don't really mind the space on disk and I believe treesitter to be
smart enough to only load the required one.
But there was one which spawned error every time I tried to run neovim.
So that was the trigger, I enabled the "auto install" option and I'll
try to pay attention to any new parser installed to decide if I want to
add it to the "ensured installed" list.

I also move the `treesitter-textobjects` extension, the `after` was not
even correct (from packer time I think... damn that's old).
Instead, I declared it as its own and put the dependency. This should
ensure it's loaded after treesitter.