nvim(lint): improve linting and formatting setup
I tried to stick to the original idea but to improve it a bit to work
with the new mono-repo setup as well as regular repositories.

One of the improvement is the usage of the `--path-mode=intersection'
option for php-cs-fixer. Both linter & formatter.
What this option does is to apply the command on the intersection of the
files provided to the command line and those discovered by the finder
setup in the configuration file.
This prevent to lint or format files that php-cs-fixer should not handle
when opening them in neovim.
Before I had some "cheats" with gloab patterns to avoid `vendor` and
`var` directory. But it wise really not great, missing specific for each
projects, etc.
So I cleaned up those checks and use the option instead.
One of the caveat is that we can't use stding for the linter anymore.
But I like better to not see "false diagnostics" and it allows to clean
up the linter config.
For the formatter it means that we have to override the template used to
construct the temporar file, we don't have a choice: when not using
stdin conform will pas through a temporary file.
And if we keep the default template, which starts with a ".", it gets
ignored by "--path-mode=intersection". Probably because it's a hidden
file.
While we could define the template with just "$FILENAME" to "use the
real file", we still see in the log a "cleanup" line so it doesn't seem
to have been thought to be used like that. I'm already used to see some
files in the git status if I looked during the run of the linter anyway.
I also redefined the "cwd" option: I would have rather like to look for
the "cwd" using the configuration file of php-cs-fixer. But since the
current mono-repo I'm working with use the same style rules on every
project: I decided to go with this for now.
Later I might make it configurable per project.

For phpstan linter, I decided to always analyze the full project.
The reason is that phpstan doesn't analyze traits included in a class
when you pass just the path to the class file. They explain why in the
documentation link I provided in the code.
Given that it takes approximatively the same time, with cache, to
analyze the full project and just one file: I decided to go with it in
order to have more accurage feedbacks.
For phpstan we don't use one global config but one for each project, so
I configured the "cwd" based on the phpstan configuration file patterns.

The rest is just a bit of cleanup